# Issue/7443 resource query further improvements