--- a/src/inmanta/data/dataview.py
+++ b/src/inmanta/data/dataview.py
@@ -53,6 +53,7 @@ from inmanta.data import (
     ResourceAction,
     ResourceHistoryOrder,
     ResourceLogOrder,
+    ResourceStatusOrder,
     SimpleQueryBuilder,
     VersionedResourceOrder,
     model,