--- a/src/inmanta/db/versions/v202403280.py
+++ b/src/inmanta/db/versions/v202403280.py
@@ -28,9 +28,10 @@ async def update(connection: Connection) -> None:
     await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN resource_type varchar")
     await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN resource_id_value varchar")
     await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN agent varchar")
-    await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_type, resource_id)")
-    await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_id_value, resource_id)")
-    await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, agent, resource_id)")
+    # TODO: these seem to not be required. Probably because scale is still relatively small when it comes to distinct resources (5000)
+    #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_type, resource_id)")
+    #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_id_value, resource_id)")
+    #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, agent, resource_id)")
     await connection.execute(
         """
         UPDATE public.resource_persistent_state AS rps