--- a/src/inmanta/db/versions/v202403280.py
+++ b/src/inmanta/db/versions/v202403280.py
@@ -20,25 +20,39 @@ from asyncpg import Connection
 
 
 async def update(connection: Connection) -> None:
-    # TODO: verify which indexes/changes are required.
-    # TODO: add UNIQUE to some indexes?
-    # TODO: cleanup
-    # TODO: document
+    """
+    Add all resource identifying columns to the resource persistent state table to improve efficiency of resource filtering and
+    sorting. These tables are derived from the resource id and are therefore part of the identity of a resource. They will not
+    change for the lifetime of a resource (with a given resource id).
+    """
+    # TODO: better still might be to have a resource identity table, then reference that one from both resource and
+    #   resource_persistent_state. Consider or follow-up ticket?
+    # TODO: migration test
 
-    await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN resource_type varchar")
-    await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN resource_id_value varchar")
-    await connection.execute("ALTER TABLE public.resource_persistent_state ADD COLUMN agent varchar")
+    await connection.execute(
+        """
+        ALTER TABLE public.resource_persistent_state
+            ADD COLUMN resource_type varchar,
+            ADD COLUMN agent varchar,
+            ADD COLUMN resource_id_value varchar
+            ;
+        UPDATE public.resource_persistent_state AS rps
+            SET (resource_type, resource_id_value, agent) = (
+                SELECT r.resource_type, resource_id_value, agent
+                FROM public.resource AS r
+                WHERE rps.resource_id = r.resource_id AND rps.environment = r.environment
+            )
+            ;
+        ALTER TABLE public.resource_persistent_state
+            ALTER COLUMN resource_type SET NOT NULL,
+            ALTER COLUMN agent SET NOT NULL,
+            ALTER COLUMN resource_id_value SET NOT NULL,
+            ADD CONSTRAINT derived_id UNIQUE (resource_type, agent, resource_id)
+            ;
+        """
+    )
+
     # TODO: these seem to not be required. Probably because scale is still relatively small when it comes to distinct resources (5000)
     #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_type, resource_id)")
     #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, resource_id_value, resource_id)")
     #await connection.execute("CREATE INDEX ON public.resource_persistent_state (environment, agent, resource_id)")
-    await connection.execute(
-        """
-        UPDATE public.resource_persistent_state AS rps
-        SET (resource_type, resource_id_value, agent) = (
-            SELECT r.resource_type, resource_id_value, agent
-            FROM public.resource AS r
-            WHERE rps.resource_id = r.resource_id AND rps.environment = r.environment
-        )
-        """
-    )