… changelogs/unreleased/7443-resource-view-query-further-improvements.yml
… tests/db/migration_tests/dumps/v202403280.sql
… tests/db/migration_tests/test_v202403120_to_v202403280.py