# Description

Made further improvements to the resource view query:
- No longer include model version in the sort key for this view: it's redundant because each resource will only be present once (and it was the final part of the sort key, after resource id). And it's theoretically expensive to calculate because it requires calculating the model version to be able to apply the filter, while otherwise the filter can be applied first, and we only calculate the model version for the rows we fetch. Impact of this change -> improves performance of unfiltered view when sorted on anything except resource status: +- 380ms -> 215ms.
- Add more resource identifying attributes (agent, resource type, resource id value) to the resource persistent state table, and return those from the query rather than the ones from resource. This allows filtering and sorting on a much smaller table before calculating the join with the resource table. Impact of this change -> greatly improves performance when filtering on these columns: 440-490ms -> 20ms.

This PR does duplicate some data between the resource and resource persistent state table. The proper relational approach would probably be to add a third table for the resource identity `(environment, resource_id, agent, resource_type, resource_id_value)`, then reference that table on `resource_id` from both the resource one and the persistent state one, and drop the derived identity attributes from it. However, this PR is an easy win, while that split would be a major effort, considering that our data framework isn't really built for it, if we want to keep the flexibility of having the derived identity fields on the Python objects.

closes #7443

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)