@wouterdb 2024-03-28
1. I wouldn't bother with overly normalizing the tables
2. if you have numbers, could you add them?