@sanderr 2024-03-28
> if you have numbers, could you add them?

I'm planning to, that's the TODO.