@sanderr 2024-03-28
With increased scale, numbers aren't as good as I'd hoped:
1. I'd hoped it would be more or less constant when scaling
2. Some filters are *a lot* better but those on status get worse with scale while they remain constant with the previous PR.

To be investigated further.